optimized gene usage
Moved from boolean[] to a new Gene class that mirrors BitSet in a lot of ways.

It is however optimized toward a fixed size given initially and has 2 vital extra functions:

copyFrom(Gene) and copyFrom(Gene, startIndex, endIndex)

which utilizes System.arraycopy for the majority of the copying. It uses Bit-maskes for the 'words' that are only getting some copied over to