Adding option to run Parallel
Added new function is AbstractPopulation:

- evolveParallel()

which runs the generation in a threadpool. ONLY available if the AbstractPopulation was initiated with 'useParallel' to true (otherwise threadpool is not initiated)

Added Islands:

New class constructed given 3 arguments (4'th optional)

It takes A convergence interface, a evolutionStep interface and lastly a list of AbstractPopulations, the optional argument is a timeout limit on threads.

Used to allow for the Island Model with GA, but can be equally useful for ACO