Last big update
- Prettified LineGraph

- Ran benchmark for OM, LO and smaller once for TSP

- Added LO fitness function (Leading Ones)

- Added 'Mutation' to Ant Colony Optimization

Melting GA and ACO together, also a clean way to add 2-opt to ACO

- Changed ACO Node graph
Major redesign to focus on Edges rather than individual arrays in nodes. Ants also hold list of edges rather than a list of 'choices' (edge-index of the node they were in). Allowing for a much more flexible altercation, as is needed for 2-opt.

- Finish GUI for framework*

*broke Islands somehow, likely some small stupid error